update main page to use all views
