--- fork-visualizer/src/app/improvedData/layout.js
@@ -1,15 +1,0 @@
-import { Inter } from 'next/font/google'
-import '@/app/globals.css'
-
-const inter = Inter({ subsets: ['latin'] })
-
-export const metadata = {
-    title: 'Create Next App',
-    description: 'Generated by create next app',
-}
-
-export default function RootLayout({ children }) {
-    return (
-        <body className={inter.className}>{children}</body>
-    )
-}
