+++ fork-visualizer/src/app/page.js
@@ -1,69 +1,64 @@
 "use client";
 
 import { ApiPromise, WsProvider } from "@polkadot/api";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import HorizontalBlockList from "@/app/components/HorizontalBlockList";
 import BlockTable from "@/app/components/BlockTable";
+import {
+    MAX_CHAIN_COUNT,
+    WS_ADDRESSES,
+    GROUP_TO_COLOR,
+    SEAL_TO_GROUP,
+    GROUP_TO_NODE_COLOR,
+    ENABLE_GRAPH_VIEW,
+    ENABLE_TEST_VIEW,
+} from "@/constants";
+import ForceGraph from "@/app/components/ForceGraph"
 
-const MAX_CHAIN_COUNT = 256;
-
-// const wsProvider = new WsProvider("ws://localhost:9944");
-const wsProvider = new WsProvider("ws://100.109.138.126:9944");
-// const wsProvider = new WsProvider("ws://localhost:9944");
 
 export default function Home() {
 
-    const [latestBlock, setLatestBlock] = useState();
-    const [data, setData] = useState({ nodes: [] });
-    const [running, setRunning] = useState(false);
+    const [ latestBlock, setLatestBlock ] = useState();
+    const [ data, setData ] = useState({ blocks: [], headers: [], nodes: [], links: [] });
+    const [ running, setRunning ] = useState(false);
+    const [ wsProviders, setWsProviders ] = useState([])
 
-    async function main() {
-        const api = await ApiPromise.create({ provider: wsProvider });
+    useEffect(() => {
+        setWsProviders([...WS_ADDRESSES.map((ws_addr) => {
+            console.log(`watching ${ws_addr}...`)
 
+            return new WsProvider(ws_addr)
+        })])
+    }, [])
 
-        // We only display a couple, then unsubscribe
-        let count = 0;
-        setRunning(true);
-        setData({ nodes: [] });
+    const updateBlocks = (header) => {
+        setLatestBlock(header.number.toHuman());
 
-        // Subscribe to the new headers on-chain. The callback is fired when new headers
-        // are found, the call itself returns a promise with a subscription that can be
-        // used to unsubscribe from the newHead subscription
-        const unsubscribe = await api.rpc.chain.subscribeNewHeads((header) => {
-            setLatestBlock(header.number.toHuman());
+        setData(({ blocks, headers, nodes, links }) => {
+            let group = "genesis";
+            let groupColor = GROUP_TO_COLOR[group]
 
-            setData(({ nodes }) => {
+            // The genesis block (number 0) does not have the normal PoW seal on it.
+            // This avoids a crash when you re-start the block. There is likely a more
+            // idiomatic way to do this is js.
+            if (header.number != 0) {
+                const seal_data = header.digest.logs[0].toJSON().seal[1];
+                const seal = seal_data.slice(0, 4)
 
-                let group = "genesis";
-                let groupColor = "bg-green-600"
+                group = SEAL_TO_GROUP[seal] ?? "genesis"
+                groupColor = GROUP_TO_COLOR[group]
+                console.log(`group: ${group}`);
+            }
 
-                // The genesis block (number 0) does not have the normal PoW seal on it.
-                // This avoids a crash when you re-start the node. There is likely a more
-                // idiomatic way to do this is js.
-                if (header.number != 0) {
-                    const seal_data = header.digest.logs[0].toJSON().seal[1];
-                    // Detect the PoW algorithm from the first byte of the seal.
-                    // This corresponds to `pub struct SupportedHashes` in multi-pow/src/lib.rs
-                    switch (seal_data.slice(0, 4)) {
-                        case "0x00":
-                            group = "md5";
-                            groupColor = "bg-red-600"
-                            break;
-                        case "0x01":
-                            group = "sha3";
-                            groupColor = "bg-blue-600"
-                            break;
-                        case "0x02":
-                            group = "keccak";
-                            groupColor = "bg-purple-600"
-                    }
+            let newBlocks = [...blocks];
 
-                    console.log(`group: ${group}`);
-                }
-
-
-                const newNodes = [
-                    ...nodes,
+            // If the block is not in the graph already, we add it.
+            // Actually it would be better to use a set datastructure here than a list,
+            // but IDK how convenient / idiomatic sets are in JS.
+            if (!blocks.some((h) => h.hash === header.hash.toString())) {
+                console.log("Block not found. Adding it to graph.")
+                newBlocks = [
+                    ...blocks,
                     {
                         hash: header.hash.toHuman(),
                         parentHash: header.parentHash.toHuman(),
@@ -74,47 +69,107 @@ export default function Home() {
                             return d.toHuman().Seal
                         }),
                         group: group,
-                        groupColor: groupColor
+                        groupColor: groupColor,
+                        duplicate: false,
                     },
                 ];
+            }
 
-                const styledNodes = newNodes.map((node, i) => {
-                    return {
-                        ...node,
-                        style:
-                            newNodes.filter((n) => n.number === node.number && n.hash !== node.hash)
-                                .length === 0
-                                ? ""
-                                : "bg-rose-900 text-white",
-                    };
-                });
-
+            const styledBlocks = newBlocks.map((block) => {
+                const duplicate = newBlocks.filter((b) => b.number === block.number && b.hash !== block.hash).length === 0
                 return {
-                    nodes: styledNodes,
+                    ...block,
+                    style: duplicate ? "" : "bg-rose-900 text-white",
+                    duplicate
                 };
             });
 
-            if (++count === MAX_CHAIN_COUNT) {
-                unsubscribe();
-                setRunning(false);
+            let newNodes = [...nodes];
+
+            // If the node is not in the graph already, we add it.
+            // Actually it would be better to use a set datastructure here than a list,
+            // but IDK how convenient / idiomatic sets are in JS.
+            if (!nodes.some((h) => h.id === header.hash.toString())) {
+                console.log("Block not found. Adding it to graph.")
+                newNodes = [
+                    ...nodes,
+                    {
+                        id: header.hash.toString(),
+                        group: group,
+                        nodeColor: GROUP_TO_NODE_COLOR[group]
+                    }
+                ];
             }
+
+            let newLinks = [...links];
+            //TODO Check out how I used `.some` above instead of `.filter`.
+            // Perhaps we should do that here as well.
+            if (newNodes.filter((h) => h.id === header.parentHash.toString()).length > 0) {
+                console.log("parent found");
+                newLinks = [
+                    ...links,
+                    {
+                        source: header.hash.toString(),
+                        target: header.parentHash.toString(),
+                    },
+                ];
+            } else {
+                console.log("parent not found");
+            }
+
+            return {
+                blocks: styledBlocks,
+                headers: [...headers, header.toHuman()],
+                nodes: newNodes,
+                links: newLinks,
+            };
         });
     }
 
+    async function main() {
+        // We only display a couple, then unsubscribe
+        let count = 0
+        setRunning(true)
+        setData({ blocks: [], headers: [], nodes: [], links: [] })
+
+        wsProviders.map(async (wsProvider) => {
+            // Subscribe to the new headers on-chain. The callback is fired when new headers
+            // are found, the call itself returns a promise with a subscription that can be
+            // used to unsubscribe from the newHead subscription
+            const api = await ApiPromise.create({ provider: wsProvider })
+            const unsubscribe = await api.rpc.chain.subscribeNewHeads((header) => {
+                updateBlocks(header)
+
+                if (++count === MAX_CHAIN_COUNT) {
+                    unsubscribe();
+                    setRunning(false);
+                }
+            });
+        })
+    }
+
     return (
-        <div className="flex flex-col justify-center h-screen overflow-clip">
-            <div className="flex items-center justify-center flex-grow gap-4">
+        <div className="flex flex-col justify-center overflow-y-scroll">
+            <div className="flex items-center justify-center flex-grow h-screen gap-4 ">
                 <div className="flex flex-col justify-center">
                     {!running && (
                         <button className="px-4 py-2 text-white bg-blue-600 rounded-full" onClick={() => main()}>start</button>
                     )}
                 </div>
-                { data.nodes.length > 0 && (
-                    <BlockTable nodes={data.nodes} latestBlock={latestBlock} />
+
+                { ENABLE_TEST_VIEW && data.blocks.length > 0 && (
+                    <BlockTable nodes={ data.blocks } latestBlock={ latestBlock } />
+                )}
+
+                { ENABLE_GRAPH_VIEW && data.blocks.length > 0 && (
+                    <ForceGraph data={ data } />
                 )}
             </div>
 
-            <HorizontalBlockList nodes={data.nodes} />
+            { ENABLE_TEST_VIEW && (
+                <HorizontalBlockList nodes={ data.blocks } />
+            )}
+
         </div>
     );
 }
